refactor: simplify error matching
