--- apps/studio/components/interfaces/ErrorHandling/errorMappings/types.ts
@@ -1,6 +1,0 @@
-import { ErrorMapping } from '../ErrorMatcher.types'
-
-export type ErrorMappingFactory = (params?: {
-  onRestartProject?: () => void
-  onDebugWithAI?: (prompt: string) => void
-}) => ErrorMapping
