+++ apps/studio/components/layouts/TableEditorLayout/TableEditorMenu.tsx
@@ -6,8 +6,7 @@ import { ExportDialog } from 'components/grid/components/header/ExportDialog'
 import { parseSupaTable } from 'components/grid/SupabaseGrid.utils'
 import { SupaTable } from 'components/grid/types'
 import { ProtectedSchemaWarning } from 'components/interfaces/Database/ProtectedSchemaWarning'
-import { MappedErrorDisplay } from 'components/interfaces/ErrorHandling/MappedErrorDisplay'
-import { SIDEBAR_KEYS } from 'components/layouts/ProjectLayout/LayoutSidebar/LayoutSidebarProvider'
+import { ErrorMatcher } from 'components/interfaces/ErrorHandling/ErrorMatcher'
 import EditorMenuListSkeleton from 'components/layouts/TableEditorLayout/EditorMenuListSkeleton'
 import { ButtonTooltip } from 'components/ui/ButtonTooltip'
 import { InfiniteListDefault, LoaderForIconMenuItems } from 'components/ui/InfiniteList'
@@ -23,8 +22,6 @@ import { useSelectedProjectQuery } from 'hooks/misc/useSelectedProject'
 import { useIsProtectedSchema } from 'hooks/useProtectedSchemas'
 import { Filter, Plus } from 'lucide-react'
 import { useCallback, useEffect, useMemo, useState } from 'react'
-import { useAiAssistantStateSnapshot } from 'state/ai-assistant-state'
-import { useSidebarManagerSnapshot } from 'state/sidebar-manager-state'
 import { useTableEditorStateSnapshot } from 'state/table-editor'
 import {
   Button,
@@ -62,8 +59,6 @@ export const TableEditorMenu = () => {
   )
 
   const { data: project } = useSelectedProjectQuery()
-  const { openSidebar } = useSidebarManagerSnapshot()
-  const aiSnap = useAiAssistantStateSnapshot()
   const {
     data,
     isLoading,
@@ -283,26 +278,18 @@ export const TableEditorMenu = () => {
 
           {/* 🚨 DELETE BEFORE MERGING */}
           {process.env.NEXT_PUBLIC_ENVIRONMENT !== 'prod' && (
-            <MappedErrorDisplay
+            <ErrorMatcher
               error="ERROR: FAILED TO RUN SQL QUERY: CONNECTION TERMINATED DUE TO CONNECTION TIMEOUT."
               supportUrl={`/support/new?project=${projectRef}`}
               className="mx-4 mt-3"
-              onDebugWithAI={(prompt) => {
-                openSidebar(SIDEBAR_KEYS.AI_ASSISTANT)
-                aiSnap.newChat({ initialMessage: prompt })
-              }}
             />
           )}
 
           {isError && (
-            <MappedErrorDisplay
+            <ErrorMatcher
               error={error?.message ?? 'Failed to retrieve tables'}
               supportUrl={`/support/new${project?.ref ? `?project=${project.ref}` : ''}`}
               className="mx-4 mt-3"
-              onDebugWithAI={(prompt) => {
-                openSidebar(SIDEBAR_KEYS.AI_ASSISTANT)
-                aiSnap.newChat({ initialMessage: prompt })
-              }}
             />
           )}
